Roll out sandbox API to stop open environments
This update introduces the functionality to utilize the sandbox API for generic stop/start operations, allowing us to manage services within environments whose contents are not known.

To activate this feature, the catalog item maintainer must configure the following in agnosticv:

```yaml
__meta__:
  sandbox_api:
    actions:
      stop:
        enable: true
      start:
        enable: true
```

When above block is present, the Sandbox API actions takes precedence over the deployer actions.

The service is stopped using the `/placements/{uuid}/stop` endpoint, and the `/requests/{request_id}/status` endpoint is monitored to confirm the completion of the operation. As this process only takes a few seconds, i think it's appropriate to maintain it within the same task file, rather than initiating another anarchy run.

Error handling is managed through the use of Ansible's block/rescue structure.

jira: GPTEINFRA-6598